Replace no-data text with prompt to refresh
