+++ projects/plugins/boost/app/assets/src/js/modules/image-size-analysis/recommendations/Hero.svelte
@@ -1,12 +1,14 @@
 <script lang="ts">
 	import { quadOut } from 'svelte/easing';
-	import { get } from 'svelte/store';
 	import { fade } from 'svelte/transition';
 	import { __ } from '@wordpress/i18n';
 	import TemplatedString from '../../../elements/TemplatedString.svelte';
 	import actionLinkTemplateVar from '../../../utils/action-link-template-var';
-	import { isaData, refreshIsaData } from '../store/isa-data';
-	import { imageDataActiveGroup, totalIssueCount } from '../store/isa-summary';
+	import { isaData } from '../store/isa-data';
+	import { imageDataActiveGroup } from '../store/isa-summary';
+
+	export let needsRefresh: boolean;
+	export let refresh: () => Promise< void >;
 
 	const formatter = new Intl.DateTimeFormat( 'en-US', {
 		month: 'long',
@@ -15,21 +17,6 @@
 		minute: 'numeric',
 		hour12: true,
 	} );
-
-	// Keep track of the total count from the summary the last time we got a data update.
-	// Useful for identify when a summary change might mean we need a refresh.
-	let countAtLastDataUpdate = 0;
-	isaData.subscribe( () => {
-		countAtLastDataUpdate = get( totalIssueCount );
-	} );
-
-	async function refresh() {
-		// Don't let the UI show a refresh button until we get fresh ISA data.
-		countAtLastDataUpdate = Infinity;
-		await refreshIsaData();
-	}
-
-	$: needsRefresh = $totalIssueCount > countAtLastDataUpdate;
 </script>
 
 {#if $imageDataActiveGroup && $isaData.data.last_updated}
